\name{extractAIC-methods}
\docType{methods}
\alias{extractAIC-methods}
\alias{extractAIC,ANY-method}
\alias{extractAIC,mle-method}
\title{ Method  \code{mle} for Function \code{extractAIC}}
\description{
 Method for  generic function \code{\link{extractAIC}} for  objects of the S4-class \code{\linkS4class{mle}} or 
\code{\linkS4class{mlePP}}.  It   is the same method as in 
\pkg{stats4} (that method is not available outside that package).}

\section{Methods}{
\describe{
\item{\code{signature(fit = "ANY")}}{}
\item{\code{signature(fit = "mle")}}{}
}}

\keyword{methods}

