\name{NHPoisson-package}
\alias{NHPoisson-package}
\alias{NHPoisson}
\docType{package}
\title{Statistical modelling of non homogeneous Poisson processes}

\description{
\pkg{NHPoisson} provides  tools for the modelling and ML estimation 
of non homogeneous Poisson processes (NHPP) in time, where the 
intensity is formulated as a function of  (time-dependent) covariates. 
A comprehensive toolkit for  
residual analysis and diagnostic of the fitted model is also provided.  Finally, it  
permits random generation of NHPP.}

\details{
\tabular{ll}{
Package: \tab NHPoisson\cr
Type: \tab Package\cr
Version: \tab 2.1\cr
Date: \tab 2014-03-05\cr
License: \tab GPL (>=2)  \cr
}}



\author{Ana C. Cebrian <acebrian@unizar.es>}

\seealso{\pkg{evir}, \pkg{extRemes}, \pkg{POT}, \pkg{ppstat}, \pkg{spatstat}}

\keyword{non homogeneous Poisson process}

